Enable dynamic search registration for plugins
